requirements.txt
Django==1.6
GitPython==1.0.1
South==1.0.2
httplib2==0.9.2
six==1.10.0
django-simple-captcha==0.5.1
django-paypal==0.2.7
django_sendmail_backend==0.1.1
pytz==2015.7
pyftpsync==1.0.3
